+++ packages/gov-ui-kit/src/core/assets/copy/coreCopy.ts
@@ -10,6 +10,7 @@ export const coreCopy = {
     dataListFilter: {
         filter: 'Filter',
         sort: 'Sort',
+        reset: 'Reset all filters',
     },
     dataListFilterStatus: {
         found: 'Found',