+++ packages/gov-ui-kit/src/core/components/dataList/dataListFilter/dataListFilterStatus.tsx
@@ -1,10 +1,16 @@
 import classNames from 'classnames';
+import { Button } from '../../button';
 import { useOdsCoreContext } from '../../odsCoreProvider';
 import { useDataListContext } from '../dataListContext';
 
-export interface IDataListFilterStatusProps {}
+export interface IDataListFilterStatusProps {
+    /**
+     * Callback to clear all active filters
+     */
+    onResetFiltersClick?: () => void;
+}
 
-export const DataListFilterStatus: React.FC<IDataListFilterStatusProps> = () => {
+export const DataListFilterStatus: React.FC<IDataListFilterStatusProps> = ({ onResetFiltersClick }) => {
     const { state, itemsCount = 0, entityLabel } = useDataListContext();
 
     const { copy } = useOdsCoreContext();
@@ -20,28 +26,35 @@ export const DataListFilterStatus: React.FC<IDataListFilterStatusProps> = () => 
     }
 
     return (
-        <p
-            className={classNames(
-                'px-3 text-sm font-normal leading-tight text-neutral-500 md:text-base',
-                { 'text-primary-400': isLoading || isInitialLoading },
-                { 'text-neutral-500': displayItemsCount },
+        <div className="flex items-center justify-between">
+            <p
+                className={classNames(
+                    'px-3 py-2.5 text-sm font-normal leading-tight text-neutral-500 md:text-base',
+                    { 'text-primary-400': isLoading || isInitialLoading },
+                    { 'text-neutral-500': displayItemsCount },
+                )}
+            >
+                {displayItemsCount && (
+                    <>
+                        <span className="text-neutral-800">{itemsCount} </span>
+                        <span>{entityLabel}</span>
+                    </>
+                )}
+                {isInitialLoading && copy.dataListFilterStatus.loadingEntity(entityLabel)}
+                {isLoading && copy.dataListFilterStatus.filteringEntity(entityLabel)}
+                {isFiltered && (
+                    <>
+                        <span>{copy.dataListFilterStatus.found} </span>
+                        <span className="text-primary-400">{itemsCount} </span>
+                        <span>{entityLabel}</span>
+                    </>
+                )}
+            </p>
+            {isFiltered && onResetFiltersClick && (
+                <Button size="sm" onClick={onResetFiltersClick} variant="ghost" responsiveSize={{ md: 'md' }}>
+                    {copy.dataListFilter.reset}
+                </Button>
             )}
-        >
-            {displayItemsCount && (
-                <>
-                    <span className="text-neutral-800">{itemsCount} </span>
-                    <span>{entityLabel}</span>
-                </>
-            )}
-            {isInitialLoading && copy.dataListFilterStatus.loadingEntity(entityLabel)}
-            {isLoading && copy.dataListFilterStatus.filteringEntity(entityLabel)}
-            {isFiltered && (
-                <>
-                    <span>{copy.dataListFilterStatus.found} </span>
-                    <span className="text-primary-400">{itemsCount} </span>
-                    <span>{entityLabel}</span>
-                </>
-            )}
-        </p>
+        </div>
     );
 };