+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   Bump `prettier-plugin-organize-imports` to 4.0.0
+
 ## [1.0.37] - 2024-07-08
 
 ### Changed