+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Add reset filter functionality to `DataListFilter` core component
+
 ### Fixed
 
 -   Remove fixed width from `EmptyState` core component